pyproject.toml
[project]
name = "marketing-tochnee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "django>=5.1.7",
    "lxml>=5.3.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
